+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ## [0.0.x]
 
 
+- Changed cyllama.pyx and tests to apply more consistent naming of Llama-type classes.
 
 ## [0.0.1]